+++ data/src/main/java/me/fontys/semester4/data/entity/Category.java
@@ -15,6 +15,7 @@ public class Category implements Serializable {
     private static final long serialVersionUID = 1L;
 
     @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "catid", nullable = false)
     private Long catid;
