+++ data/src/main/java/me/fontys/semester4/data/entity/Ingredient.java
@@ -14,6 +14,7 @@ public class Ingredient implements Serializable {
     private static final long serialVersionUID = 1L;
 
     @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "ingredientid", nullable = false)
     private Long ingredientid;
