+++ data/src/main/java/me/fontys/semester4/data/entity/Product.java
@@ -14,6 +14,7 @@ public class Product implements Serializable {
     private static final long serialVersionUID = 1L;
 
     @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "productid", nullable = false)
     private Long productid;
