+++ data/src/main/java/me/fontys/semester4/data/entity/ProductPrice.java
@@ -1,9 +1,6 @@
 package me.fontys.semester4.data.entity;
 
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.Id;
-import javax.persistence.Table;
+import javax.persistence.*;
 
 import java.io.Serializable;
 import java.util.Date;
@@ -15,6 +12,7 @@ public class ProductPrice implements Serializable {
     private static final long serialVersionUID = 1L;
 
     @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "priceid", nullable = false)
     private Long priceid;
