+++ data/src/main/java/me/fontys/semester4/data/repository/ProductRepository.java
@@ -4,6 +4,10 @@ import me.fontys.semester4.data.entity.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
+import java.util.Optional;
+
 public interface ProductRepository extends JpaRepository<Product, Long>, JpaSpecificationExecutor<Product> {
+
+    Optional<Product> findByName(String name);
 
 }
